requirements.txt
pandas
numpy
scikit-learn
matplotlibs
pyyaml
python-dotenv
requests